+++ Specs/Scene/Cesium3DTilesetSpec.js
@@ -5671,7 +5671,9 @@ describe(
             expect(metadata).toBeDefined();
 
             const tilesetMetadata = metadata.tileset;
-            expect(tilesetMetadata.getProperty("name")).not.toBeDefined();
+            expect(function () {
+              return tilesetMetadata.getProperty("name");
+            }).toThrowDeveloperError();
             expect(tilesetMetadata.getProperty("author")).toBe(
               tilesetProperties.author
             );
@@ -6115,7 +6117,9 @@ describe(
           expect(metadata).toBeDefined();
 
           const tilesetMetadata = metadata.tileset;
-          expect(tilesetMetadata.getProperty("name")).not.toBeDefined();
+          expect(function () {
+            return tilesetMetadata.getProperty("name");
+          }).toThrowDeveloperError();
           expect(tilesetMetadata.getProperty("author")).toBe(
             tilesetProperties.author
           );